refactor: separate `dataclasses` & rename `parser`->`reader`
