+++ karriz/mp3_file_reader/error.py
@@ -2,10 +2,6 @@ class FileStreamNotLoadedError(Exception):
     def __init__(self):
         super().__init__('MP3 파일이 로드 되지않았습니다.')
 
-class FileStreamNotLoadedError(Exception):
-    def __init__(self):
-        super().__init__('MP3 파일이 로드 되지않았습니다.')
-                
 class InvalidFrameSyncError(Exception):
     def __init__(self):
         super().__init__('유효하지 않은 FrameSync 값 입니다.')
